% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motdat.R
\docType{data}
\name{motdat}
\alias{motdat}
\title{Example data used in Furr & Heuckeroth, on Impression Motivation (IM)}
\format{A data frame of 90 rows and 13 variables:
 \describe{
   \item{imscale}{Impression Motivation}
   \item{dep}{Dependence}
   \item{mach}{Machiavellianism}
   \item{dis}{Distrust}
   \item{res}{Resourcefulness}
   \item{se}{Self-efficacy}
   \item{ext}{Extraversion}
   \item{agr}{Agreeableness}
   \item{comp}{Complexity}
   \item{psc}{Public self-consciousness}
   \item{sm}{Self-monitoring}
   \item{anx}{Anxiety}
   \item{nfb}{Need to belong}
   }}
\usage{
data(motdat)
}
\description{
Self-report data from 90 participants, responding to the IM scale and
12 criterion variables.
}
\examples{
data(motdat)
}
\keyword{datasets}
