% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.qcv.R
\name{print.qcv}
\alias{print.qcv}
\title{Printing Quantifying Construct Validity}
\usage{
\method{print}{qcv}(x, ...)
}
\arguments{
\item{x}{object of class "qcv" (i.e., output from the qcv() function}

\item{...}{More arguments to pass to the print function.}
}
\description{
Prints key results from the qcv() function
}
\examples{
actrIM  <- c(.46, .13, -.24, -.03, .12, .03, .39, .06, .51, .08, .24, .66)
predrIM <- c(.58, .24, -.04, .06, -.04, .18, .36, .08, .64, .56, .36, .56)
imqcvout <- qcv(n=90, actr=actrIM, predr=predrIM, medr=.075)
print(imqcvout)
}
